requirements.txt
Django==2.0.4
pbkdf2==1.3
pycryptodome==3.6.1
pycryptodomex==3.6.1
pytz==2018.4
